int => decimal(10,2)

+++ CoCowork.DB/Tables/LaptopOrder.sql
@@ -5,7 +5,7 @@
 	[OrderId] int NOT NULL,
 	[StartDate] datetime NOT NULL,
 	[EndDate] datetime NOT NULL,
-	[SubtotalPrice] int NOT NULL,
+	[SubtotalPrice] decimal(10,2) NOT NULL,
 	CONSTRAINT [FK_LaptopId_to_Laptop] FOREIGN KEY ([LaptopId]) REFERENCES [Laptop]([Id]), 
     CONSTRAINT [FK_LaptopOrder_Order] FOREIGN KEY ([OrderId]) REFERENCES [Order]([Id])
 )
